+++ src/main/webapp/WEB-INF/jsp/comments.jsp
@@ -4,7 +4,7 @@
 <!DOCTYPE html>
 
 <html lang="en">
-<title>Hello world</title>
+<title>Comment section</title>
 <%-- Links to Bootstrap CSS framework as a webjar dependency --%>
 <link rel="stylesheet" type="text/css"
       href="webjars/bootstrap/3.3.5/css/bootstrap.min.css"/>
@@ -16,21 +16,23 @@
 |   |   |  \/ /_/  >  \        /\  ___/| \_\ \  |   |  /    / 
 |___|___|  /\___  /    \__/\  /  \___  >___  /  |___| /____/  
          \//_____/          \/       \/    \/                 
-<!-- A free, customizable, responsive weather widget: https://weatherwidget.io/ -->
-<a class="weatherwidget-io" href="https://forecast7.com/en/41d65n0d89/zaragoza/" data-label_1="ZARAGOZA" data-label_2="WEATHER" data-theme="original" >ZARAGOZA WEATHER</a>
-<script>
-!function(d,s,id){var js,fjs=d.getElementsByTagName(s)[0];if(!d.getElementById(id)){js=d.createElement(s);js.id=id;js.src='https://weatherwidget.io/js/widget.min.js';fjs.parentNode.insertBefore(js,fjs);}}(document,'script','weatherwidget-io-js');
-</script>
+<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 </pre>
-<kbd>${os} ${hostname} ${version} ${time} <br>
-Last login: ${last_time} from ${last_ip}<br>
-<br>Deadline is ${deadline}<br>
-${daysLeft} days, ${hoursLeft} hours, ${minutesLeft} minutes and ${secondsLeft} seconds left<br>
-<br>
-Do you wanna know if you would survive thanos snap?? ->> <a href="./thanos">CLICK HERE</a><br>
-<br>If you want to write a comment or view the comments, check the <a href="./comments">comment section</a>.<br>
-<br>
-user@${hostname}:~$ ${message}</kbd>
+<kbd>
+
+<h2><font color=white>Comments</font></h2>
+<c:forEach items="${comments}" var="comment">
+      <c:out value="${comment}<br>" escapeXml="false"/><br>
+</c:forEach>
+
+</kbd>
+
+<form action="/comments" method=POST> <font color=white>Write a comment!</font> <br>
+    <input type="text" name="comment" placeholder="Comment"><br><br>
+    <input type="text" name="name" placeholder="Your name"><br><br>
+    <input type="submit" value="Submit">
+</form>
+
 <%-- Links to JQuery JavaScript library as a webjar dependency --%>
 <script type="text/javascript" src="webjars/jquery/2.1.4/jquery.min.js"></script>
 </body>
